v_3.40
v_3.40

+++ Work/Common/Include/rnd/Perlin.h
@@ -293,6 +293,6 @@ public:
     }
 };
 
-using Perlin = PerlinNoise<double>;
+using Perlin = PerlinNoise<float>;
 
 } //namespace rnd
